Search SDK: Combining DocumentsOperations source files
