--- src/Search/Search/Customizations/Documents/DocumentsOperationsExtensions.Suggest.Customization.cs
@@ -1,185 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for
-// license information.
-
-namespace Microsoft.Azure.Search
-{
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Search.Models;
-    using Microsoft.Rest.Azure;
-
-    public static partial class DocumentsOperationsExtensions
-    {
-        /// <summary>
-        /// Suggests query terms based on input text and matching documents in the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798936.aspx"/>
-        /// </summary>
-        /// <param name='operations'>
-        /// The operations group for this extension method.
-        /// </param>
-        /// <param name="searchText">
-        /// The search text on which to base suggestions.
-        /// </param>
-        /// <param name="suggesterName">
-        /// The name of the suggester as specified in the suggesters collection that's part of the index definition.
-        /// </param>
-        /// <param name="suggestParameters">
-        /// Parameters to further refine the suggestion query.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <returns>
-        /// Response containing the suggested text and documents matching the query.
-        /// </returns>
-        /// <remarks>
-        /// The non-generic overloads of the Suggest, SuggestAsync, and SuggestWithHttpMessagesAsync methods make a
-        /// best-effort attempt to map JSON types in the response payload to .NET types. See
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more information.
-        /// </remarks>
-        public static DocumentSuggestResult Suggest(
-            this IDocumentsOperations operations,
-            string searchText,
-            string suggesterName,
-            SuggestParameters suggestParameters = null,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions))
-        {
-            return Task.Factory.StartNew(s => ((IDocumentsOperations)s).SuggestAsync(searchText, suggesterName, suggestParameters, searchRequestOptions), operations, CancellationToken.None, TaskCreationOptions.None, TaskScheduler.Default).Unwrap().GetAwaiter().GetResult();
-        }
-
-        /// <summary>
-        /// Suggests query terms based on input text and matching documents in the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798936.aspx"/>
-        /// </summary>
-        /// <param name='operations'>
-        /// The operations group for this extension method.
-        /// </param>
-        /// <param name="searchText">
-        /// The search text on which to base suggestions.
-        /// </param>
-        /// <param name="suggesterName">
-        /// The name of the suggester as specified in the suggesters collection that's part of the index definition.
-        /// </param>
-        /// <param name="suggestParameters">
-        /// Parameters to further refine the suggestion query.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <param name='cancellationToken'>
-        /// The cancellation token.
-        /// </param>
-        /// <returns>
-        /// Response containing the suggested text and documents matching the query.
-        /// </returns>
-        /// <remarks>
-        /// The non-generic overloads of the Suggest, SuggestAsync, and SuggestWithHttpMessagesAsync methods make a
-        /// best-effort attempt to map JSON types in the response payload to .NET types. See
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more information.
-        /// </remarks>
-        public static async Task<DocumentSuggestResult> SuggestAsync(
-            this IDocumentsOperations operations,
-            string searchText,
-            string suggesterName,
-            SuggestParameters suggestParameters = null,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions),
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            AzureOperationResponse<DocumentSuggestResult> result = await operations.SuggestWithHttpMessagesAsync(searchText, suggesterName, suggestParameters ?? new SuggestParameters(), searchRequestOptions, null, cancellationToken).ConfigureAwait(false);
-            return result.Body;
-        }
-
-        /// <summary>
-        /// Suggests query terms based on input text and matching documents in the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798936.aspx"/>
-        /// </summary>
-        /// <typeparam name="T">
-        /// The CLR type that maps to the index schema. Instances of this type can be retrieved as documents
-        /// from the index.
-        /// </typeparam>
-        /// <param name='operations'>
-        /// The operations group for this extension method.
-        /// </param>
-        /// <param name="searchText">
-        /// The search text on which to base suggestions.
-        /// </param>
-        /// <param name="suggesterName">
-        /// The name of the suggester as specified in the suggesters collection that's part of the index definition.
-        /// </param>
-        /// <param name="suggestParameters">
-        /// Parameters to further refine the suggestion query.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <returns>
-        /// Response containing the suggested text and documents matching the query.
-        /// </returns>
-        /// <remarks>
-        /// The generic overloads of the Suggest, SuggestAsync, and SuggestWithHttpMessagesAsync methods support
-        /// mapping of Azure Search field types to .NET types via the type parameter T. See 
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync&lt;T&gt;(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more details on the type mapping.
-        /// </remarks>
-        public static DocumentSuggestResult<T> Suggest<T>(
-            this IDocumentsOperations operations,
-            string searchText,
-            string suggesterName,
-            SuggestParameters suggestParameters = null,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions))
-            where T : class
-        {
-            return Task.Factory.StartNew(s => ((IDocumentsOperations)s).SuggestAsync<T>(searchText, suggesterName, suggestParameters, searchRequestOptions), operations, CancellationToken.None, TaskCreationOptions.None, TaskScheduler.Default).Unwrap().GetAwaiter().GetResult();
-        }
-
-        /// <summary>
-        /// Suggests query terms based on input text and matching documents in the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798936.aspx"/>
-        /// </summary>
-        /// <typeparam name="T">
-        /// The CLR type that maps to the index schema. Instances of this type can be retrieved as documents
-        /// from the index.
-        /// </typeparam>
-        /// <param name='operations'>
-        /// The operations group for this extension method.
-        /// </param>
-        /// <param name="searchText">
-        /// The search text on which to base suggestions.
-        /// </param>
-        /// <param name="suggesterName">
-        /// The name of the suggester as specified in the suggesters collection that's part of the index definition.
-        /// </param>
-        /// <param name="suggestParameters">
-        /// Parameters to further refine the suggestion query.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <param name='cancellationToken'>
-        /// The cancellation token.
-        /// </param>
-        /// <returns>
-        /// Response containing the suggested text and documents matching the query.
-        /// </returns>
-        /// <remarks>
-        /// The generic overloads of the Suggest, SuggestAsync, and SuggestWithHttpMessagesAsync methods support
-        /// mapping of Azure Search field types to .NET types via the type parameter T. See 
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync&lt;T&gt;(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more details on the type mapping.
-        /// </remarks>
-        public static async Task<DocumentSuggestResult<T>> SuggestAsync<T>(
-            this IDocumentsOperations operations,
-            string searchText,
-            string suggesterName,
-            SuggestParameters suggestParameters = null,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions),
-            CancellationToken cancellationToken = default(CancellationToken)) where T : class
-        {
-            AzureOperationResponse<DocumentSuggestResult<T>> result = await operations.SuggestWithHttpMessagesAsync<T>(searchText, suggesterName, suggestParameters ?? new SuggestParameters(), searchRequestOptions, null, cancellationToken).ConfigureAwait(false);
-            return result.Body;
-        }
-    }
-}
