--- src/Search/Search/Customizations/Documents/IDocumentsOperations.Index.Customization.cs
@@ -1,96 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for
-// license information.
-
-namespace Microsoft.Azure.Search
-{
-    using System.Collections.Generic;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Search.Models;
-    using Microsoft.Rest.Azure;
-
-    public partial interface IDocumentsOperations
-    {
-        /// <summary>
-        /// Sends a batch of upload, merge, and/or delete actions to the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798935.aspx"/>
-        /// </summary>
-        /// <param name="batch">
-        /// The batch of index actions.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <param name='customHeaders'>
-        /// The headers that will be added to request.
-        /// </param>
-        /// <param name='cancellationToken'>
-        /// The cancellation token.
-        /// </param>
-        /// <exception cref="IndexBatchException">
-        /// Thrown when some of the indexing actions failed, but other actions succeeded and modified the state of
-        /// the index. This can happen when the Search Service is under heavy indexing load. It is important to
-        /// explicitly catch this exception and check its
-        /// <c cref="IndexBatchException.IndexingResults">IndexResult</c> property. This property reports the status
-        /// of each indexing action in the batch, making it possible to determine the state of the index after a
-        /// partial failure.
-        /// </exception>
-        /// <remarks>
-        /// The non-generic overloads of the Index, IndexAsync, and IndexWithHttpMessagesAsync methods make a
-        /// best-effort attempt to map JSON types in the response payload to .NET types. See
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more information.
-        /// </remarks>
-        /// <returns>
-        /// Response containing the status of operations for all actions in the batch.
-        /// </returns>
-        Task<AzureOperationResponse<DocumentIndexResult>> IndexWithHttpMessagesAsync(
-            IndexBatch batch,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions),
-            Dictionary<string, List<string>> customHeaders = null,
-            CancellationToken cancellationToken = default(CancellationToken));
-
-        /// <summary>
-        /// Sends a batch of upload, merge, and/or delete actions to the Azure Search index.
-        /// <see href="https://msdn.microsoft.com/library/azure/dn798935.aspx"/>
-        /// </summary>
-        /// <typeparam name="T">
-        /// The CLR type that maps to the index schema. Instances of this type can be stored as documents in the index.
-        /// </typeparam>
-        /// <param name="batch">
-        /// The batch of index actions.
-        /// </param>
-        /// <param name='searchRequestOptions'>
-        /// Additional parameters for the operation
-        /// </param>
-        /// <param name='customHeaders'>
-        /// The headers that will be added to request.
-        /// </param>
-        /// <param name='cancellationToken'>
-        /// The cancellation token.
-        /// </param>
-        /// <exception cref="IndexBatchException">
-        /// Thrown when some of the indexing actions failed, but other actions succeeded and modified the state of
-        /// the index. This can happen when the Search Service is under heavy indexing load. It is important to
-        /// explicitly catch this exception and check its
-        /// <c cref="IndexBatchException.IndexingResults">IndexResult</c> property. This property reports the status
-        /// of each indexing action in the batch, making it possible to determine the state of the index after a
-        /// partial failure.
-        /// </exception>
-        /// <returns>
-        /// Response containing the status of operations for all actions in the batch.
-        /// </returns>
-        /// <remarks>
-        /// The generic overloads of the Index, IndexAsync, and IndexWithHttpMessagesAsync methods support mapping of
-        /// Azure Search field types to .NET types via the type parameter T. See 
-        /// <see cref="IDocumentsOperations.GetWithHttpMessagesAsync&lt;T&gt;(string, System.Collections.Generic.IEnumerable&lt;string&gt;, SearchRequestOptions, System.Collections.Generic.Dictionary&lt;string, System.Collections.Generic.List&lt;string&gt;&gt;, CancellationToken)"/>
-        /// for more details on the type mapping.
-        /// </remarks>
-        Task<AzureOperationResponse<DocumentIndexResult>> IndexWithHttpMessagesAsync<T>(
-            IndexBatch<T> batch,
-            SearchRequestOptions searchRequestOptions = default(SearchRequestOptions),
-            Dictionary<string, List<string>> customHeaders = null,
-            CancellationToken cancellationToken = default(CancellationToken)) where T : class;
-    }
-}
